+++ src/lib.rs
@@ -107,11 +107,11 @@ pub use ifo::{
     AudioApplicationMode, AudioAttributes, AudioCodingMode, AudioLanguageType,
     AudioQuantizationDrc, CellAddrEntry, CellPlaybackInfo, CellPositionInfo, DvdChapter, DvdTitle,
     DvdTitleEntry, FrameRate, McExtensionEntry, MenuAttributes, NavCommand, PaletteEntry, Pgc,
-    PgcCommandTable, PgcTime, Pgci, PgciSrp, Ptt, PttTitle, SubpictureAttributes,
+    PgcCommandTable, PgcTime, Pgci, PgciSrp, PtlMait, Ptt, PttTitle, SubpictureAttributes,
     SubpictureCodingMode, SubpictureLanguageType, TitleAttributes, TmapEntry, TtSrpt,
     VideoAspectRatio, VideoAttributes, VideoCodingMode, VideoResolution, VideoStandard, VmgIfo,
-    VobuAdmap, VtsCAdt, VtsIfo, VtsPttSrpt, VtsTmap, VtsTmapti, VtsiMat, DVD_SECTOR, VMG_MAGIC,
-    VTS_MAGIC,
+    VmgPtlMait, VmgVtsAtrt, VmgVtsAtrtEntry, VobuAdmap, VtsCAdt, VtsIfo, VtsPttSrpt, VtsTmap,
+    VtsTmapti, VtsiMat, DVD_SECTOR, VMG_MAGIC, VTS_MAGIC,
 };
 pub use iso9660::{
     DirectoryRecord, Iso9660Entry, Iso9660Volume, PathTableEntry, PrimaryVolumeDescriptor,